+++ module-api/src/main/java/com/whiskey/auth/controller/AuthController.java
@@ -1,8 +1,8 @@
 package com.whiskey.auth.controller;
 
 import com.whiskey.auth.dto.LoginRequest;
+import com.whiskey.auth.service.AuthService;
 import com.whiskey.domain.auth.JwtResponse;
-import com.whiskey.member.service.MemberService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -16,11 +16,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final MemberService memberService;
+    private final AuthService authService;
 
     @PostMapping("/login")
     public ResponseEntity<JwtResponse> login(@Valid @RequestBody LoginRequest loginRequest) {
-        JwtResponse response = memberService.login(loginRequest);
+        JwtResponse response = authService.login(loginRequest);
         return ResponseEntity.ok(response);
     }
 }
