+++ module-api/src/main/java/com/whiskey/member/service/MemberService.java
@@ -29,7 +29,6 @@ public class MemberService {
 
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
-    private final JwtTokenProvider jwtTokenProvider;
 
     @Transactional
     public void signup(MemberRegisterValue memberDto) {
@@ -59,42 +58,6 @@ public class MemberService {
             ValidationErrorValue errorValue = new ValidationErrorValue("email", email, "duplicate");
             throw ErrorCode.CONFLICT.exception("이미 가입된 이메일입니다.", errorValue);
         }
-    }
-
-    public JwtResponse login(@Valid LoginRequest loginRequest) {
-        try {
-            Member member = authenticateMember(loginRequest.email(), loginRequest.password());
-
-            List<SimpleGrantedAuthority> authorities = List.of(
-                new SimpleGrantedAuthority("ROLE_USER")
-            );
-
-            String accessToken = jwtTokenProvider.generateToken(member.getId(), authorities);
-            String refreshToken = jwtTokenProvider.generateRefreshToken(member.getId());
-
-            Long expireTime = jwtTokenProvider.getAccessTokenValidityTime();
-            MemberInfo memberInfo = MemberInfo.from(member);
-
-            return new JwtResponse(accessToken, refreshToken, "Bearer", expireTime, memberInfo);
-        }
-        catch(Exception e) {
-            throw ErrorCode.UNAUTHORIZED.exception("인증에 실패했습니다.");
-        }
-    }
-
-    private Member authenticateMember(String email, String password) {
-        Member member = memberRepository.findByEmailAndStatus(email, MemberStatus.ACTIVE)
-            .orElseThrow(() -> ErrorCode.UNAUTHORIZED.exception("존재하지 않는 회원입니다."));
-
-        if(!passwordEncoder.matches(password, member.getPasswordHash())) {
-            throw ErrorCode.UNAUTHORIZED.exception("비밀번호가 일치하지 않습니다.");
-        }
-
-        if(member.getStatus() != MemberStatus.ACTIVE) {
-            throw ErrorCode.UNAUTHORIZED.exception("비활성화된 계정입니다.");
-        }
-
-        return member;
     }
 
     public MemberResponse getMemberById(Long id) {
